Add logging aspect
